--- widgets/waterfall/components/add-to-cart/base.template.ts
@@ -1,5 +1,0 @@
-export default () => {
-  return `
-    <div id="stacklapopup-add-to-cart"></div>
-  `;
-};
