+++ widgets/waterfall/widget.ts
@@ -21,6 +21,7 @@ import shopspotStyle from "./components/shopspot-icon/base.scss";
 import customExpandedTileTemplate from "./components/expanded-tile/base.template.hbs";
 import getCSSVariables from "widgets/libs/css-variables";
 import { getTimephrase } from "widgets/libs/tile.lib";
+import { expandedTileView } from "./components/expanded-tile/expanded-tile.view";
 
 declare const sdk: ISdkMasonry;
 
@@ -52,17 +53,4 @@ sdk.addCSSToComponent(expandedTileCSS, "expanded-tile");
 sdk.addCSSToComponent(productsCSS, "ugc-products");
 sdk.addCSSToComponent(shopspotStyle, "shopspot-icon");
 
-sdk.addTemplateToComponent(customExpandedTileTemplate, async () => {
-  const tile = sdk.tiles.getTile();
-  return {
-    shopspotEnabled: sdk.isComponentLoaded('shopspots'),
-    productsEnabled: sdk.isComponentLoaded('products'),
-    parent: sdk.getNodeId() || '',
-    tile: sdk.tiles.getTile(),
-    showTimestamp: tile && tile.source_created_at && widgetSettings.expanded_tile_show_timestamp,
-    timephrase: tile && getTimephrase(tile.source_created_at),
-    showCaption: tile && tile.message && widgetSettings.expanded_tile_show_caption,
-    showSharing: widgetSettings.expanded_tile_show_sharing
-  }
-},
-"expanded-tile");
+sdk.addTemplateToComponent(customExpandedTileTemplate, expandedTileView, "expanded-tile");
